Move `openInteractiveDialog` logic to its own module
The integration_actions module was being loaded implicitly by another
part of the code, and it stopped being loaded in
70bae6af74546cfb1849b659a0cce23cbd3ed353.

This moves it into its own module, pending a deeper refactoring.
